Fix scripts create 211225

+++ rr-et/app/Http/Controllers/ScriptController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Category;
 use App\Script;
 use Illuminate\Http\Request;
 use App\Http\Requests\CreateScript;
 use App\Http\Requests\EditScript;
+use Illuminate\Support\Facades\Auth;
 
 class ScriptController extends Controller
 {
@@ -46,7 +48,9 @@ class ScriptController extends Controller
      */
     public function create()
     {
-        return view('scripts/create');
+        $categories = Category::all();
+
+        return view('scripts/create', compact('categories'));
     }
 
     /**
@@ -57,8 +61,14 @@ class ScriptController extends Controller
      */
     public function store(CreateScript $request)
     {
+        $user = Auth::user();
+        $userId = $user->id;
+
         $script = new Script();
+
         $script->content = $request->content;
+        $script->user_id = $userId;
+        $script->category_id = $request->category_id;
         $script->save();
 
         return redirect()->route('scripts.index');
@@ -89,9 +99,14 @@ class ScriptController extends Controller
     {
         $script = Script::find($id);
 
-        return view('scripts/edit', [
-            'script' => $script,
-        ]);
+        $loggedInUser = Auth::user();
+        if ($script->user_id !== $loggedInUser->id) {
+            return redirect()->route('scripts.index');
+        }
+
+        $categories = Category::all();
+
+        return view('scripts/edit', compact('script', 'categories'));
     }
 
     /**
@@ -101,11 +116,17 @@ class ScriptController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(EditScript $request, $id)
+    public function update(EditScript $request, int $id)
     {
         $script = Script::find($id);
 
+        $loggedInUser = Auth::user();
+        if ($script->user_id !== $loggedInUser->id) {
+            return redirect()->route('scripts.index');
+        }
+
         $script->content = $request->content;
+        $script->category_id = $request->category_id;
         $script->save();
 
         return redirect()->route('scripts.index');
@@ -120,6 +141,11 @@ class ScriptController extends Controller
     public function destroy(int $id)
     {
         $script = Script::find($id);
+
+        $loggedInUser = Auth::user();
+        if ($script->user_id !== $loggedInUser->id) {
+            return redirect()->route('scripts.index');
+        }
 
         $script->delete();
         return redirect(route('scripts.index'));
